# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/books": {
      "post": {
        "operationId": "BooksController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Success"
                    },
                    "data": {
                      "oneOf": [
                        {
                          "$ref": "#/components/schemas/Book"
                        },
                        {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Book"
                          }
                        }
                      ],
                      "nullable": true
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Token tidak ada / tidak valid",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Token tidak ada / tidak valid"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "UNAUTHORIZED"
                        },
                        "details": {
                          "oneOf": [
                            {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "field": {
                                    "type": "string",
                                    "example": "email"
                                  },
                                  "message": {
                                    "type": "string",
                                    "example": "Email sudah terdaftar"
                                  }
                                }
                              }
                            },
                            {
                              "type": "null"
                            }
                          ],
                          "example": null
                        }
                      }
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "error"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Role tidak diizinkan",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Role tidak diizinkan"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "FORBIDDEN"
                        },
                        "details": {
                          "oneOf": [
                            {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "field": {
                                    "type": "string",
                                    "example": "email"
                                  },
                                  "message": {
                                    "type": "string",
                                    "example": "Email sudah terdaftar"
                                  }
                                }
                              }
                            },
                            {
                              "type": "null"
                            }
                          ],
                          "example": null
                        }
                      }
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "error"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Tambah buku baru",
        "tags": [
          "Books"
        ]
      },
      "get": {
        "operationId": "BooksController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Kata kunci pencarian",
            "schema": {
              "example": "search term",
              "type": "string"
            }
          },
          {
            "name": "action",
            "required": false,
            "in": "query",
            "description": "Filter berdasarkan jenis aksi audit log",
            "schema": {
              "example": "CREATE",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Success"
                    },
                    "data": {
                      "oneOf": [
                        {
                          "$ref": "#/components/schemas/Book"
                        },
                        {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Book"
                          }
                        }
                      ],
                      "nullable": true
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "number",
                          "example": 1
                        },
                        "limit": {
                          "type": "number",
                          "example": 10
                        },
                        "total_items": {
                          "type": "number",
                          "example": 0
                        },
                        "total_pages": {
                          "type": "number",
                          "example": 0
                        },
                        "has_next_page": {
                          "type": "boolean",
                          "example": false
                        },
                        "has_prev_page": {
                          "type": "boolean",
                          "example": false
                        }
                      }
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "summary": "List semua buku (paginated)",
        "tags": [
          "Books"
        ]
      }
    },
...
    "/books/items": {
      "post": {
        "operationId": "BooksController_createItem",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Success"
                    },
                    "data": {
                      "oneOf": [
                        {
                          "$ref": "#/components/schemas/BookItem"
                        },
                        {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/BookItem"
                          }
                        }
                      ],
                      "nullable": true
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Token tidak ada / tidak valid",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Token tidak ada / tidak valid"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "UNAUTHORIZED"
                        },
                        "details": {
                          "oneOf": [
                            {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "field": {
                                    "type": "string",
                                    "example": "email"
                                  },
                                  "message": {
                                    "type": "string",
                                    "example": "Email sudah terdaftar"
                                  }
                                }
                              }
                            },
                            {
                              "type": "null"
                            }
                          ],
                          "example": null
                        }
                      }
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "error"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Role tidak diizinkan",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Role tidak diizinkan"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "FORBIDDEN"
                        },
                        "details": {
                          "oneOf": [
                            {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "field": {
                                    "type": "string",
                                    "example": "email"
                                  },
                                  "message": {
                                    "type": "string",
                                    "example": "Email sudah terdaftar"
                                  }
                                }
                              }
                            },
                            {
                              "type": "null"
                            }
                          ],
                          "example": null
                        }
                      }
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "error"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Tambah eksemplar fisik buku",
        "tags": [
          "Books"
        ]
      }
    },
    "/books/{id}/items": {
      "get": {
        "operationId": "BooksController_findAllItems",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Success"
                    },
                    "data": {
                      "type": "object",
                      "nullable": true
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "data"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Buku tidak ditemukan",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Buku tidak ditemukan"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "NOT_FOUND"
                        },
                        "details": {
                          "oneOf": [
                            {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "field": {
                                    "type": "string",
                                    "example": "email"
                                  },
                                  "message": {
                                    "type": "string",
                                    "example": "Email sudah terdaftar"
                                  }
                                }
                              }
                            },
                            {
                              "type": "null"
                            }
                          ],
                          "example": null
                        }
                      }
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "error"
                  ]
                }
              }
            }
          }
        },
        "summary": "List eksemplar sebuah buku",
        "tags": [
          "Books"
        ]
      }
    },
    "/books/bulk": {
      "post": {
        "operationId": "BooksController_createMany",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Success"
                    },
                    "data": {
                      "oneOf": [
                        {
                          "$ref": "#/components/schemas/Book"
                        },
                        {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Book"
                          }
                        }
                      ],
                      "nullable": true
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Token tidak ada / tidak valid",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Token tidak ada / tidak valid"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "UNAUTHORIZED"
                        },
                        "details": {
                          "oneOf": [
                            {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "field": {
                                    "type": "string",
                                    "example": "email"
                                  },
                                  "message": {
                                    "type": "string",
                                    "example": "Email sudah terdaftar"
                                  }
                                }
                              }
                            },
                            {
                              "type": "null"
                            }
                          ],
                          "example": null
                        }
                      }
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "error"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Role tidak diizinkan",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Role tidak diizinkan"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "FORBIDDEN"
                        },
                        "details": {
                          "oneOf": [
                            {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "field": {
                                    "type": "string",
                                    "example": "email"
                                  },
                                  "message": {
                                    "type": "string",
                                    "example": "Email sudah terdaftar"
                                  }
                                }
                              }
                            },
                            {
                              "type": "null"
                            }
                          ],
                          "example": null
                        }
                      }
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "error"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Tambah banyak buku sekaligus",
        "tags": [
          "Books"
        ]
      }
    },
...
  },
  "info": {
    "title": "Library Digital API",
...
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Alternatif untuk cookie: tempel access_token hasil login."
      },
      "access_token": {
        "type": "http",
        "in": "cookie",
        "name": "access_token",
        "description": "HttpOnly cookie access token, diset otomatis oleh POST /auth/login."
      }
    },
    "schemas": {
...
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "books": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Book"
            }
          }
        },
        "required": [
          "id",
          "name",
          "books"
        ]
      },
      "Publisher": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "books": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Book"
            }
          }
        },
        "required": [
          "id",
          "name",
          "address",
          "books"
        ]
      },
...
      "BookItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "barcode": {
            "type": "string"
          },
          "status": {
            "enum": [
              "AVAILABLE",
              "RESERVED",
              "BORROWED",
              "LOST",
              "DAMAGED"
            ],
            "type": "string"
          },
          "condition": {
            "enum": [
              "GOOD",
              "SLIGHTLY_DAMAGED",
              "HEAVILY_DAMAGED"
            ],
            "type": "string"
          },
          "book": {
            "$ref": "#/components/schemas/Book"
          },
          "added_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "barcode",
          "status",
          "condition",
          "book",
          "added_at",
          "updated_at"
        ]
      },
      "Book": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "sub_title": {
            "type": "string"
          },
          "isbn_13": {
            "type": "string"
          },
          "isbn_10": {
            "type": "string"
          },
          "published_year": {
            "type": "number"
          },
          "description": {
            "type": "string"
          },
          "image_url": {
            "type": "string"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "publisher": {
            "$ref": "#/components/schemas/Publisher"
          },
          "language": {
            "$ref": "#/components/schemas/Language"
          },
          "authors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Author"
            }
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookItem"
            }
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "sub_title",
          "isbn_13",
          "isbn_10",
          "published_year",
          "description",
          "image_url",
          "category",
          "publisher",
          "language",
          "authors",
          "items",
          "created_at",
          "updated_at"
        ]
      },
...
      "CreateBookDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Clean Code"
          },
          "sub_title": {
            "type": "string",
            "example": "A Handbook of Agile Software Craftsmanship"
          },
          "isbn_13": {
            "type": "string",
            "example": "9780132350884"
          },
          "isbn_10": {
            "type": "string",
            "example": "0132350882"
          },
          "published_year": {
            "type": "number",
            "example": 2008
          },
          "description": {
            "type": "string",
            "example": "A book about writing clean code"
          },
          "category_id": {
            "type": "number",
            "example": 1,
            "description": "ID kategori buku"
          },
          "publisher_id": {
            "type": "number",
            "example": 1,
            "description": "ID penerbit"
          },
          "language_id": {
            "type": "number",
            "example": 1,
            "description": "ID bahasa"
          },
          "author_ids": {
            "example": [
              1,
              2
            ],
            "description": "Array ID penulis",
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "title",
          "category_id",
          "publisher_id",
          "language_id",
          "author_ids"
        ]
      },
...
      "CreateBookItemDto": {
        "type": "object",
        "properties": {
          "barcode": {
            "type": "string",
            "example": "BRC-001",
            "description": "Barcode unik eksemplar"
          },
          "book_id": {
            "type": "number",
            "example": 1,
            "description": "ID buku induk"
          },
          "status": {
            "enum": [
              "AVAILABLE",
              "RESERVED",
              "BORROWED",
              "LOST",
              "DAMAGED"
            ],
            "type": "string",
            "example": "AVAILABLE"
          },
          "condition": {
            "enum": [
              "GOOD",
              "SLIGHTLY_DAMAGED",
              "HEAVILY_DAMAGED"
            ],
            "type": "string",
            "example": "GOOD"
          }
        },
        "required": [
          "barcode",
          "book_id"
        ]
      },
...
    }
  }
}